requirements.txt
flask == 2.2.5
flask_restx == 1.3.0
mongoengine == 0.29.1
flask_mongoengine == 1.0.0
gunicorn==20.1.0
python-dotenv
google-genai == 1.9.0
pydantic